destroy! now once again calls destroy
This is because destroy! is called by :dependent => :destroy association definitions (yes, destroy! and not destroy), which means that associated records of a paranoid object would suddenly go missing... making it very hard to restore them!

So instead, it's now called really_destroy! and destroy! is just an alias for destroy once again