Remove rubinius failing build nonsense
I am tired of supporting Rubinius. I don't use it, nor know of anyone who uses it. Every damn month it breaks on Travis in some new and brilliant way and I don't have the time or energy to support it.

If people truly care about it, then *they* can support it.