Fix for 'cannot touch a new record object' error.
This is a fix for a case when a model has a has_one association with an
after_validation callback, and we have failed validation. Paranoia will
attempt to touch the deleted_at column on the child record that has
been built but not saved. See pull request for more information.