Add ability to have types other than datetime
This allows an optional :type option to be passed to acts_as_paranoid.
The allowed types are :datetime (the default), and :boolean. If the
column type is boolean, true will be written to the attribute rather
than the date/time.

It should follow all the same behavior as the datetime version, except
for callbacks. The default implementation uses AR::Base#touch which
fires after_touch, after_commit, and after_rollback callbacks. The
boolean version uses update_column which fires no call backs. I
couldn't find an equivalent method for a non-datetime field.