Ensure deleted_at and updated_at share the same (ruby obj) value on delete
Previously, on a delete call, `current_time_from_proper_timezone` was called
twice, which would work in most scenarios but in fact leads to two distinct
values whose difference can grow for a variety of reasons. The solution is
to memoize the value and share it between the two columns on save.