Use GitHub Action instead of travis-ci.org
CI doesn't work against recent PRs because building on travis-ci.org is ceased.

We can use travis-ci.com instead of travis-ci.org, but it has limitations to build time. So I think using GitHub Actions is the best CI to build this repository.

I migrated the settings while maintaining the original state as much as possible. Improvements to the configuration, such as adding Rails 7.0, will be done in the next PR.